Replace 128 bit integer with ptr-width integer, as it should never reach beyond 32 or even 64 bits of entropy, and save on casting efforts for native arch

+++ src/binary/leb128.rs
@@ -3,15 +3,15 @@ use crate::binary::error::ParseError;
 // Yields length in buffer and value
 #[inline]
 pub fn uleb128(bytes: &[u8]) -> Result<(usize, usize), ParseError> {
-	let mut result = 0u128;
+	let mut result = 0_usize;
 	let mask = 1 << 7;
 
 	for (i, current) in bytes.iter().enumerate() {
-		let bits = (current & (mask - 1)) as u128;
+		let bits = (current & (mask - 1)) as usize;
 		result |=  bits << (7 * i);
 
 		if mask & current == 0 {
-			return Ok((i + 1, result as usize));
+			return Ok((i + 1, result));
 		};
 	}
 	if bytes.len() == 0 {
